% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cccd_classifier.R
\name{cccd_classifier}
\alias{cccd_classifier}
\title{Class Cover Catch Digraph Classifier}
\usage{
cccd_classifier(x, y, proportion = 1)
}
\arguments{
\item{x}{asd.}

\item{y}{asd}

\item{proportion}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Class Cover Catch Digraph Classification.
}
\details{
asd
}
\examples{

rnorm(100)

}
\references{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
