% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcccd_ensemble_classifier.R
\name{pcccd_ensemble_classifier}
\alias{pcccd_ensemble_classifier}
\title{Pure and Proper Class Cover Catch Digraph Ensemble Classifier}
\usage{
pcccd_ensemble_classifier(
  x,
  y,
  n_model = 30,
  n_var = ncol(x),
  replace = FALSE,
  prop_sample = ifelse(replace, 1, 0.67),
  min_proportion = 0.7,
  max_proportion = 1,
  verbose = TRUE
)
}
\arguments{
\item{x}{feature matrix or dataframe.}

\item{y}{class factor variable.}

\item{n_model}{an integer. Number of weak classifiers.}

\item{n_var}{an integer. number of variables in weak classifiers.}

\item{replace}{a bool. Should replacement be used in data sampling}

\item{prop_sample}{a value between 0 and 1. Proportion the number of resampled
samples to the number of samples in x.}

\item{min_proportion}{Minimum proportion of cover proportion in weak classifiers.}

\item{max_proportion}{Maximum proportion of cover proportion in weak classifiers.}
}
\value{
an object of "cccd_classifier" which includes:
 \item{i_dominant_list}{dominant sample indexes.}
 \item{x_dominant_list}{dominant samples from feature matrix, x}
 \item{radii_dominant_list}{Radiuses of the circle for dominant samples}
 \item{class_names}{class names}
 \item{k_class}{number of classes}
 \item{proportions}{proportions each class covered}
}
\description{
\code{pcccd_ensemble_classifier} fits an Ensemble Pure and Proper Class Cover Catch
Digraph (PCCCD) classification model.
}
\details{
Bagging framework for PCCCD.
}
\examples{
n <- 1000

}
\author{
Fatih Saglam, saglamf89@gmail.com
}
