% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cccd_classifier.R
\name{predict.cccd_classifier}
\alias{predict.cccd_classifier}
\title{Class Cover Catch Digraph Predictor}
\usage{
\method{predict}{cccd_classifier}(object, newdata, type = "pred", ...)
}
\arguments{
\item{object}{asd.}

\item{newdata}{asd}

\item{type}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Class Cover Catch Digraph Prediction
}
\details{
asd
}
\examples{

rnorm(100)

}
\references{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
