% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.pcccd_ensemble_classifier.R
\name{predict.pcccd_ensemble_classifier}
\alias{predict.pcccd_ensemble_classifier}
\title{Pure and Proper Class Cover Catch Digraph Ensemble Prediction}
\usage{
\method{predict}{pcccd_ensemble_classifier}(object, newdata, type = "pred", ...)
}
\arguments{
\item{object}{a \code{rwcccd_classifier} object}

\item{newdata}{newdata as matrix or dataframe.}

\item{type}{"pred" or "prob". Default is "pred". "pred" is class estimations,
"prob" is \eqn{n\times k} matrix of class probabilities.}

\item{...}{not used.}
}
\value{
a vector of class predictions (if type is "pred") or a \eqn{n\times p}
matrix of class probabilities (if type is "prob").
}
\description{
\code{predict.pcccd_ensemble_classifier} makes prediction using
\code{pcccd_ensemble_classifier} object.
}
\examples{
n <- 1000

}
\author{
Fatih Saglam, saglamf89@gmail.com
}
