% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rwcccd_classifier.R
\name{predict.rwcccd_classifier}
\alias{predict.rwcccd_classifier}
\title{Random Walk Class Cover Catch Digraph Predictor}
\usage{
\method{predict}{rwcccd_classifier}(object, newdata, type = "pred", e = 0, ...)
}
\arguments{
\item{object}{asd.}

\item{newdata}{asd}

\item{type}{asd}

\item{e}{asd}

\item{...}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Random Walk Class Cover Catch Digraph Prediction
}
\details{
asd
}
\examples{

rnorm(100)

}
\references{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
