% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcccd_classifier.R
\name{pcccd_classifier}
\alias{pcccd_classifier}
\title{Pure and Proper Class Cover Catch Digraph Classifier}
\usage{
pcccd_classifier(x, y, proportion = 1)
}
\arguments{
\item{x}{feature matrix or dataframe.}

\item{y}{class factor variable.}

\item{proportion}{proportion of covered samples. A real number between \eqn{(-1,1)}.
1 by default. Smaller numbers results in less dominant samples.}
}
\value{
an object of "cccd_classifier" which includes:
 \item{i_dominant_list}{dominant sample indexes.}
 \item{x_dominant_list}{dominant samples from feature matrix, x}
 \item{radii_dominant_list}{Radiuses of the circle for dominant samples}
 \item{class_names}{class names}
 \item{k_class}{number of classes}
 \item{proportions}{proportions each class covered}
}
\description{
Pure and Proper Class Cover Catch Digraph (PCCCD) Classification.
}
\details{
Multiclass framework for pure and proper class cover catch digraph (PCCCD).
PCCCD determines target class dominant points set \eqn{S} and their circular cover area by determining
balls \eqn{B(x^{\text{target}}, r_{i})} with radii r using minimum amount of
dominant point which satisfies \eqn{X^{\text{non-target}}\cap \bigcup_{i}
B_{i} = \varnothing} (pure) and \eqn{X^{\text{target}}\subset \bigcup_{i}
B_{i}} (proper).

This guarantees that balls of target class never covers any non-target
samples (pure) and balls cover all target samples (proper). For detail,
please refer to Priebe et al. (2001), Marchette and Socolinsky (2003), and
Manukyan and Ceyan (2016).

Note: Much faster than `cccd` package.
}
\examples{
n <- 1000
x1 <- runif(n, 1, 10)
x2 <- runif(n, 1, 10)
x <- cbind(x1, x2)
y <- as.factor(ifelse(3 < x1 & x1 < 7 & 3 < x2 & x2 < 7, "A", "B"))

m_pcccd <- pcccd_classifier(x = x, y = y)

# dataset
plot(x, col = y, asp = 1)

# dominant samples of first class
x_center <- m_pcccd$x_dominant_list[[1]]

# radii of balls for first class
radii <- m_pcccd$radii_dominant_list[[1]]

# balls
for (i in 1:nrow(x_center)) {
xx <- x_center[i, 1]
yy <- x_center[i, 2]
r <- radii[i]
theta <- seq(0, 2*pi, length.out = 100)
xx <- xx + r*cos(theta)
yy <- yy + r*sin(theta)
lines(xx, yy, type = "l", col = "green")
}

# testing the performance
i_train <- sample(1:n, round(n*0.8))

x_train <- x[i_train,]
y_train <- y[i_train]

x_test <- x[-i_train,]
y_test <- y[-i_train]

m_pcccd <- pcccd_classifier(x = x_train, y = y_train)
pred <- predict(object = m_pcccd, newdata = x_test)

# confusion matrix
table(y_test, pred)

# test accuracy
sum(y_test == pred)/nrow(x_test)

}
\references{
Priebe, C. E., DeVinney, J. G., & Marchette, D. J. (2001). On the distribution of the domination number for random class cover catch digraphs. Statistics & Probability Letters, 55(3), 239-246.

Marchette, C. E. P. D. J., & Socolinsky, J. G. D. D. A. (2003). Classiﬁcation Using Class Cover Catch Digraphs. Journal of Classiﬁcation, 20, 3-23.

Manukyan, A., & Ceyhan, E. (2016). Classification of imbalanced data with a geometric digraph family. The Journal of Machine Learning Research, 17(1), 6504-6543.
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
