% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwcccd_classifier.R
\name{rwcccd_classifier}
\alias{rwcccd_classifier}
\title{Random Walk Class Cover Catch Digraph Classifier}
\usage{
rwcccd_classifier(x, y, method = "a", m = 1, proportion = 0.99)
}
\arguments{
\item{x}{feature matrix or dataframe.}

\item{y}{class factor variable.}

\item{method}{asd}

\item{m}{asd}

\item{proportion}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Random Walk Class Cover Catch Digraph Classification.
}
\details{
asd
}
\examples{

rnorm(100)

}
\references{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
